스플래시, 심볼 관련 함수 extension 파일로 분리
